Remove misleading content from index.html
index.html is misleading and many think that the content in here affects how the application look. The code is also wrong (imports example-template but uses hello-world). I removed all rows that do not affect the app in any way, and added a comment describing the purpose of the file. 

This patch was created with an iPhone.